+++ R/plot_implicit_forecast.R
@@ -38,7 +38,7 @@ implicit_forecasts_plot.default <- function(
 		n_last_tc = 4,
 		...,
 		sa = NULL, i_f = NULL
-		){
+){
 	tc <- object
 	i_f <- ts(c(sa[length(sa)], i_f),
 			  end = end(i_f),
@@ -50,7 +50,7 @@ implicit_forecasts_plot.default <- function(
 	colnames(complete_data)[2] <- c("implicit forecast")
 
 	if (is.null(ylim) & !is.null(xlim))
-		ylim <- range(window(complete_data, start = xlim[1], end = xlim[2], extend = TRUE), na.rm = TRUE)
+		ylim <- get_ylim(complete_data, xlim)
 
 	plot(complete_data, type = "l", plot.type = "single",
 		 xlim = xlim, ylim = ylim,
@@ -70,7 +70,8 @@ implicit_forecasts_plot.tc_estimates <- function(
 		lty_i_f = 3,
 		n_last_tc = 4,
 		...){
-
+	if (is.null(n_last_tc))
+		n_last_tc <- mcd(object)
 	implicit_forecasts_plot.default(
 		sa = object[["x"]],
 		i_f = implicit_forecasts(object),
@@ -123,21 +124,26 @@ ggimplicit_forecasts_plot.default <- function(
 	complete_data <- ts.union(sa, i_f, tc, tc_final, tc_prov)
 
 	if (is.null(ylim) & !is.null(xlim))
-		ylim <- range(window(complete_data, start = xlim[1], end = xlim[2], extend = TRUE), na.rm = TRUE)
+		ylim <- get_ylim(complete_data, xlim)
 
 	data <- data.frame(time = as.numeric(time(complete_data)),
 					   complete_data,
 					   check.names = FALSE)
 
-	ggplot2::ggplot(data = data, ggplot2::aes(x = time)) +
+	p <- ggplot2::ggplot(data = data, ggplot2::aes(x = time)) +
 		ggplot2::geom_line(ggplot2::aes(y = sa, color = legend_sa), na.rm = TRUE) +
 		ggplot2::geom_line(ggplot2::aes(y = i_f, color = legend_i_f), lty = lty_i_f, na.rm = TRUE) +
-		ggplot2::geom_line(ggplot2::aes(y = tc_final, color = legend_tc), na.rm = TRUE) +
-		ggplot2::geom_line(ggplot2::aes(y = tc_prov), color = col_tc, lty = lty_last_tc, na.rm = TRUE) +
+		ggplot2::geom_line(ggplot2::aes(y = tc_final, color = legend_tc), na.rm = TRUE)
+	if (n_last_tc > 0)
+		p <- p +
+		ggplot2::geom_line(ggplot2::aes(y = tc_prov), color = col_tc, lty = lty_last_tc, na.rm = TRUE)
+	p <- p  +
 		ggplot2::scale_color_manual(values = c(col_sa, col_i_f, col_tc)) +
 		ggplot2::theme(legend.title = ggplot2::element_blank()) +
-		ggplot2::labs(x = NULL, y = NULL) +
-		ggplot2::coord_cartesian(xlim = xlim, ylim = ylim)
+		ggplot2::labs(x = NULL, y = NULL)
+	if (!is.null(xlim) | !is.null(ylim))
+		p <- p + ggplot2::coord_cartesian(xlim = xlim, ylim = ylim)
+	p
 }
 #' @export
 ggimplicit_forecasts_plot.tc_estimates <- function(
@@ -152,6 +158,8 @@ ggimplicit_forecasts_plot.tc_estimates <- function(
 		legend_sa = "Seasonally adjusted",
 		legend_i_f = "Implicit forecasts",
 		...){
+	if (is.null(n_last_tc))
+		n_last_tc <- mcd(object)
 	ggimplicit_forecasts_plot.default(
 		sa = object[["x"]],
 		i_f = implicit_forecasts(object),