+++ NEWS.md
@@ -1,3 +1,19 @@
+# publishTC 0.2.2
+
+- New function `x13_regarima_outliers()` to detect AO and LS on a seasonally adjusted series.
+
+- New function `smoothness()` to compute a smoothness statistic.
+
+- If `n_last_tc = NULL` then the value is defined according to the MCD statistic.
+
+- `ggconfint_plot()` correction of parameters `xlim` and `ylim` which were not used. 
+
+- New parameter `sa_bar_line` for `growthplot()`.
+
+# publishTC 0.2.1
+
+- Improvement of `ggplot2` plots.
+
 # publishTC 0.2.0
 
 - `mcd()` now returns by default the value of the number of period (instead of `NULL`)